requirements.txt
aiogram
aiosqlite
beautifulsoup4
fpdf2
lxml
opencv-python-headless
python-dotenv
ujson

